вертикальные повороты кривые. надо переделать

+++ includes/math_utilits.h
@@ -6,7 +6,7 @@
 float			point_side(float px, float py, t_xy vert, t_xy vert1);
 t_xy			intersection_point(t_fline wall, t_fline twall);
 int				determine_intersection(t_fline wall, t_fline twall);
-int				clamp(int a, int min, int max);
+float			clamp(float a, float min, float max);
 int				int_max(int a, int b);
 int				int_min(int a, int b);
 float			vector_cross_product(float x0, float y0, float x1, float y1);
